backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pillow==11.0.0
numpy==2.1.3
python-multipart==0.0.20

# Model + training-time deps used at inference (OpenCLIP + sklearn)
open_clip_torch==2.26.1
torch>=2.2.0
torchvision>=0.17.0
scikit-learn==1.5.2
joblib==1.4.2